pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "combine-cmyk"
version = "0.0.5"
description = "Combine CMYK channel TIFFs into a single CMYK TIFF"
authors = [
  { name="David Guerrro", email="guerrero.david2@gmail.com" }
]
dependencies = [
  "pillow",
  "tifffile"
]

